refactor(storage/connect): strip pgxpool helper, expose only BuildIAMAuthToken
The pgxpool config helper introduced earlier in this PR had no in-tree
consumer in go-libs; the only consumer is ledger-v3-poc#577 (mirror v2
Postgres source). Move every pgxpool-side concern (config builder, hook
chaining, IAM auth option, integration test) back to ledger and keep go-libs'
contribution to the single primitive only ledger needs: BuildIAMAuthToken,
a thin OTel-aware wrapper around auth.BuildAuthToken.

This makes the go-libs PR strictly additive:
- iam.go is restored to its origin/main state (no refactor of iamConnector)
- new BuildIAMAuthToken lives in iam_token.go; no existing call site is
  rewired to use it -- callers in production (banking-bridge, portal,
  console via iamConnector) are untouched
- the only test is the SigV4 structural one on BuildIAMAuthToken itself

If a future consumer needs the pgxpool plumbing in go-libs, we add it then
with a real call site in mind.
